Destroy associations async respect query constraints
The codepaths related to destroying associations asynchronously now
consider when query constraints are present. In most cases, this means
interpreting the primary key as an array of columns, and identifying
associated records by a tuple of these columns, where previously this
would've been a single ID. In each of the callsites, we use branching.
This is done to maintain backwards compatibility and ensure the
signature of the destroy job remains stable: it has consumers outside of
Rails.

+++ activerecord/lib/active_record/destroy_association_async_job.rb
@@ -23,8 +23,15 @@ module ActiveRecord
         raise DestroyAssociationAsyncError, "owner record not destroyed"
       end
 
-      association_model.where(association_primary_key_column => association_ids).find_each do |r|
-        r.destroy
+      if association_model.query_constraints_list
+        association_ids
+          .map { |assoc_ids| association_model.where(association_primary_key_column.zip(assoc_ids).to_h) }
+          .inject(&:or)
+          .find_each { |r| r.destroy }
+      else
+        association_model.where(association_primary_key_column => association_ids).find_each do |r|
+          r.destroy
+        end
       end
     end
